@@ -72,22 +72,22 @@ Tickets::Application.configure do
   # Send deprecation notices to registered listeners.
   config.active_support.deprecation = :notify
 
-  config.action_mailer.default_url_options = { :host => '21ticket.ru' }
+  config.action_mailer.default_url_options = { host: '21ticket.ru' }
   # ActionMailer Config
   # Setup for production - deliveries, no errors raised
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.perform_deliveries = true
   config.action_mailer.raise_delivery_errors = false
-  config.action_mailer.default :charset => "utf-8"
+  config.action_mailer.default charset: "utf-8"
 
   config.action_mailer.smtp_settings = {
-    :address => "smtp.yandex.ru",
-    :port => 465,
-    :domain => "21ticket.ru",
-    :authentication => :plain,
-    :user_name => ENV["GMAIL_USERNAME"],
-    :password => ENV["GMAIL_PASSWORD"],
-    :tls => true
+    address: "smtp.yandex.ru",
+    port: 465,
+    domain: "21ticket.ru",
+    authentication: :plain,
+    user_name: ENV["GMAIL_USERNAME"],
+    password: ENV["GMAIL_PASSWORD"],
+    tls: true
   }
 
 
